+++ src/importlinter/contracts/layers.py
@@ -1,18 +1,18 @@
-import copy
-from typing import Iterator, List, Optional, Set, Tuple, Union, cast
+from __future__ import annotations
 
-from grimp import DetailedImport
+import itertools
+from typing import List, cast
+
+import grimp
 from typing_extensions import TypedDict
 
 from importlinter.application import contract_utils, output
 from importlinter.application.contract_utils import AlertLevel
-from importlinter.configuration import settings
 from importlinter.domain import fields
 from importlinter.domain.contract import Contract, ContractCheck, InvalidContractOptions
 from importlinter.domain.imports import Module
-from grimp import ImportGraph
 
-from ._common import DetailedChain, find_segments, render_chain_data, segments_to_collapsed_chains
+from ._common import Chain, DetailedChain, Link, render_chain_data
 
 
 class Layer:
@@ -22,7 +22,7 @@ class Layer:
 
 
 class LayerField(fields.Field):
-    def parse(self, raw_data: Union[str, List]) -> Layer:
+    def parse(self, raw_data: str | list) -> Layer:
         raw_string = fields.StringField().parse(raw_data)
         if raw_string.startswith("(") and raw_string.endswith(")"):
             layer_name = raw_string[1:-1]
@@ -36,7 +36,7 @@ class LayerField(fields.Field):
 class _LayerChainData(TypedDict):
     higher_layer: str
     lower_layer: str
-    chains: List[DetailedChain]
+    chains: list[DetailedChain]
 
 
 _UNKNOWN_LINE_NUMBER = -1
@@ -90,10 +90,7 @@ class LayersContract(Contract):
                 }
             )
 
-    def check(self, graph: ImportGraph, verbose: bool) -> ContractCheck:
-        is_kept = True
-        invalid_chains = []
-
+    def check(self, graph: grimp.ImportGraph, verbose: bool) -> ContractCheck:
         warnings = contract_utils.remove_ignored_imports(
             graph=graph,
             ignore_imports=self.ignore_imports,  # type: ignore
@@ -106,42 +103,20 @@ class LayersContract(Contract):
             self._check_all_containerless_layers_exist(graph)
 
         undeclared_modules = self._get_undeclared_modules(graph)
-        if undeclared_modules:
-            is_kept = False
 
-        for (
-            higher_layer_package,
-            lower_layer_package,
-            container,
-        ) in self._generate_module_permutations(graph):
-            output.verbose_print(
-                verbose,
-                "Searching for import chains from "
-                f"{lower_layer_package} to {higher_layer_package}...",
-            )
-            with settings.TIMER as timer:
-                layer_chain_data = self._build_layer_chain_data(
-                    higher_layer_package=higher_layer_package,
-                    lower_layer_package=lower_layer_package,
-                    container=container,
-                    graph=graph,
-                )
-
-            if layer_chain_data["chains"]:
-                is_kept = False
-                invalid_chains.append(layer_chain_data)
-            if verbose:
-                chain_count = len(layer_chain_data["chains"])
-                pluralized = "s" if chain_count != 1 else ""
-                output.print(
-                    f"Found {chain_count} illegal chain{pluralized} "
-                    f"in {timer.duration_in_s}s.",
-                )
+        dependencies = graph.find_illegal_dependencies_for_layers(
+            layers=tuple(layer.name for layer in self.layers),  # type: ignore
+            containers=self.containers,  # type: ignore
+        )
+        invalid_chains = self._build_invalid_chains(dependencies, graph)
 
         return ContractCheck(
-            kept=is_kept,
+            kept=not (dependencies or undeclared_modules),
             warnings=warnings,
-            metadata={"invalid_chains": invalid_chains, "undeclared_modules": undeclared_modules},
+            metadata={
+                "invalid_chains": invalid_chains,
+                "undeclared_modules": undeclared_modules,
+            },
         )
 
     def render_broken_contract(self, check: ContractCheck) -> None:
@@ -168,7 +143,7 @@ class LayersContract(Contract):
             )
             output.new_line()
 
-    def _validate_containers(self, graph: ImportGraph) -> None:
+    def _validate_containers(self, graph: grimp.ImportGraph) -> None:
         root_package_names = self.session_options["root_packages"]
         root_packages = tuple(Module(name) for name in root_package_names)
 
@@ -192,7 +167,9 @@ class LayersContract(Contract):
                 raise ValueError(error_message)
             self._check_all_layers_exist_for_container(container, graph)
 
-    def _check_all_layers_exist_for_container(self, container: str, graph: ImportGraph) -> None:
+    def _check_all_layers_exist_for_container(
+        self, container: str, graph: grimp.ImportGraph
+    ) -> None:
         for layer in self.layers:  # type: ignore
             if layer.is_optional:
                 continue
@@ -203,14 +180,14 @@ class LayersContract(Contract):
                     f"module {layer_module_name} does not exist."
                 )
 
-    def _get_undeclared_modules(self, graph: ImportGraph) -> Set[str]:
+    def _get_undeclared_modules(self, graph: grimp.ImportGraph) -> set[str]:
         if not self.exhaustive:
             return set()
 
         undeclared_modules = set()
 
-        exhaustive_ignores: Set[str] = self.exhaustive_ignores or set()  # type: ignore
-        layers: Set[str] = {layer.name for layer in self.layers}  # type: ignore
+        exhaustive_ignores: set[str] = self.exhaustive_ignores or set()  # type: ignore
+        layers: set[str] = {layer.name for layer in self.layers}  # type: ignore
         declared_modules = layers | exhaustive_ignores
 
         for container in self.containers:  # type: ignore[attr-defined]
@@ -221,7 +198,7 @@ class LayersContract(Contract):
 
         return undeclared_modules
 
-    def _check_all_containerless_layers_exist(self, graph: ImportGraph) -> None:
+    def _check_all_containerless_layers_exist(self, graph: grimp.ImportGraph) -> None:
         for layer in self.layers:  # type: ignore
             if layer.is_optional:
                 continue
@@ -230,182 +207,84 @@ class LayersContract(Contract):
                     f"Missing layer '{layer.name}': module {layer.name} does not exist."
                 )
 
-    def _generate_module_permutations(
-        self, graph: ImportGraph
-    ) -> Iterator[Tuple[Module, Module, Optional[str]]]:
-        """
-        Return all possible combinations of higher level and lower level modules, in pairs.
-
-        Each pair of modules consists of immediate children of two different layers. The first
-        module is in a layer higher than the layer of the second module. This means the first
-        module is allowed to import the second, but not the other way around.
-
-        Returns:
-            module_in_higher_layer, module_in_lower_layer, container
-        """
-        # If there are no containers, we still want to run the loop once.
-        quasi_containers = self.containers or [None]  # type: ignore
-
-        for container in quasi_containers:  # type: ignore
-            for index, higher_layer in enumerate(self.layers):  # type: ignore
-                higher_layer_module = self._module_from_layer(higher_layer, container)
-
-                if higher_layer_module.name not in graph.modules:
-                    continue
-
-                for lower_layer in self.layers[index + 1 :]:  # type: ignore
-
-                    lower_layer_module = self._module_from_layer(lower_layer, container)
-
-                    if lower_layer_module.name not in graph.modules:
-                        continue
-
-                    yield higher_layer_module, lower_layer_module, container
-
-    def _module_from_layer(self, layer: Layer, container: Optional[str] = None) -> Module:
+    def _module_from_layer(self, layer: Layer, container: str | None = None) -> Module:
         if container:
             name = ".".join([container, layer.name])
         else:
             name = layer.name
         return Module(name)
 
-    def _build_layer_chain_data(
-        self,
-        higher_layer_package: Module,
-        lower_layer_package: Module,
-        container: Optional[str],
-        graph: ImportGraph,
-    ) -> _LayerChainData:
-        """
-        Build a dictionary of illegal chains between two layers, in the form:
-
-            higher_layer (str): Higher layer package name.
-            lower_layer (str):  Lower layer package name.
-            chains (list):      List of <detailed chain> lists.
-        """
-        temp_graph = copy.deepcopy(graph)
-        self._remove_other_layers(
-            temp_graph,
-            container=container,
-            layers_to_preserve=(higher_layer_package, lower_layer_package),
-        )
-        # Assemble direct imports between the layers, then remove them.
-        import_details_between_layers = self._pop_direct_imports(
-            higher_layer_package=higher_layer_package,
-            lower_layer_package=lower_layer_package,
-            graph=temp_graph,
-        )
-
-        collapsed_direct_chains: List[DetailedChain] = []
-        for import_details_list in import_details_between_layers:
-            line_numbers = tuple(
-                j["line_number"] if j["line_number"] != _UNKNOWN_LINE_NUMBER else None
-                for j in import_details_list
-            )
-            collapsed_direct_chains.append(
-                {
-                    "chain": [
-                        {
-                            "importer": import_details_list[0]["importer"],
-                            "imported": import_details_list[0]["imported"],
-                            "line_numbers": line_numbers,
-                        }
-                    ],
-                    "extra_firsts": [],
-                    "extra_lasts": [],
-                }
-            )
-
-        layer_chain_data: _LayerChainData = {
-            "higher_layer": higher_layer_package.name,
-            "lower_layer": lower_layer_package.name,
-            "chains": collapsed_direct_chains,
-        }
-
-        indirect_chain_data = self._get_indirect_collapsed_chains(
-            temp_graph, importer_package=lower_layer_package, imported_package=higher_layer_package
-        )
-        layer_chain_data["chains"].extend(indirect_chain_data)
-
-        return layer_chain_data
-
-    @classmethod
-    def _get_indirect_collapsed_chains(
-        cls, graph: ImportGraph, importer_package: Module, imported_package: Module
-    ) -> List[DetailedChain]:
-        """
-        Squashes the two packages.
-        Gets a list of paths between them, called segments.
-        Add the heads and tails to the segments.
-        Return a list of detailed chains in the following format:
-
-        [
+    def _build_invalid_chains(
+        self, dependencies: set[grimp.PackageDependency], graph: grimp.ImportGraph
+    ) -> list[_LayerChainData]:
+        return [
             {
-                "chain": <detailed chain>,
-                "extra_firsts": [
-                    <import details>,
-                    ...
-                ],
-                "extra_lasts": [
-                    <import details>,
-                    <import details>,
-                    ...
+                "higher_layer": dependency.downstream,
+                "lower_layer": dependency.upstream,
+                "chains": [
+                    self._build_detailed_chain_from_route(c, graph) for c in dependency.routes
                 ],
             }
+            for dependency in dependencies
         ]
-        """
-        temp_graph = copy.deepcopy(graph)
 
-        temp_graph.squash_module(importer_package.name)
-        temp_graph.squash_module(imported_package.name)
+    def _build_detailed_chain_from_route(
+        self, route: grimp.Route, graph: grimp.ImportGraph
+    ) -> DetailedChain:
+        ordered_heads = sorted(route.heads)
+        extra_firsts: list[Link] = [
+            {
+                "importer": head,
+                "imported": route.middle[0],
+                "line_numbers": self._get_line_numbers(
+                    importer=head, imported=route.middle[0], graph=graph
+                ),
+            }
+            for head in ordered_heads[1:]
+        ]
+        ordered_tails = sorted(route.tails)
+        extra_lasts: list[Link] = [
+            {
+                "imported": tail,
+                "importer": route.middle[-1],
+                "line_numbers": self._get_line_numbers(
+                    imported=tail, importer=route.middle[-1], graph=graph
+                ),
+            }
+            for tail in ordered_tails[1:]
+        ]
+        chain_as_strings = [ordered_heads[0], *route.middle, ordered_tails[0]]
+        chain_as_links: Chain = [
+            {
+                "importer": importer,
+                "imported": imported,
+                "line_numbers": self._get_line_numbers(
+                    importer=importer, imported=imported, graph=graph
+                ),
+            }
+            for importer, imported in _pairwise(chain_as_strings)
+        ]
+        return {
+            "chain": chain_as_links,
+            "extra_firsts": extra_firsts,
+            "extra_lasts": extra_lasts,
+        }
 
-        segments = find_segments(
-            temp_graph, reference_graph=graph, importer=importer_package, imported=imported_package
-        )
-        return segments_to_collapsed_chains(
-            graph, segments, importer=importer_package, imported=imported_package
-        )
+    def _get_line_numbers(
+        self, importer: str, imported: str, graph: grimp.ImportGraph
+    ) -> tuple[int | None, ...]:
+        details = graph.get_import_details(importer=importer, imported=imported)
+        line_numbers = tuple(i["line_number"] for i in details) if details else (None,)
+        return line_numbers
 
-    def _remove_other_layers(self, graph: ImportGraph, container, layers_to_preserve):
-        for index, layer in enumerate(self.layers):  # type: ignore
-            candidate_layer = self._module_from_layer(layer, container)
-            if candidate_layer.name in graph.modules and candidate_layer not in layers_to_preserve:
-                self._remove_layer(graph, layer_package=candidate_layer)
 
-    def _remove_layer(self, graph: ImportGraph, layer_package):
-        for module in graph.find_descendants(layer_package.name):
-            graph.remove_module(module)
-        graph.remove_module(layer_package.name)
+def _pairwise(iterable):
+    """
+    Return successive overlapping pairs taken from the input iterable.
+    pairwise('ABCDEFG') --> AB BC CD DE EF FG
 
-    @classmethod
-    def _pop_direct_imports(
-        cls, higher_layer_package, lower_layer_package, graph: ImportGraph
-    ) -> List[List[DetailedImport]]:
-        import_details_list = []
-        lower_layer_modules = {lower_layer_package.name} | graph.find_descendants(
-            lower_layer_package.name
-        )
-        for lower_layer_module in lower_layer_modules:
-            imported_modules = graph.find_modules_directly_imported_by(lower_layer_module).copy()
-            for imported_module in imported_modules:
-                if Module(imported_module) == higher_layer_package or Module(
-                    imported_module
-                ).is_descendant_of(higher_layer_package):
-                    import_details = graph.get_import_details(
-                        importer=lower_layer_module, imported=imported_module
-                    )
-                    if not import_details:
-                        # get_import_details may not return any imports (for example if an import
-                        # has been added without metadata. If nothing is returned, we still add
-                        # details, to keep the type checker happy.
-                        import_details = [
-                            {
-                                "importer": lower_layer_module,
-                                "imported": imported_module,
-                                "line_number": _UNKNOWN_LINE_NUMBER,
-                                "line_contents": "",
-                            }
-                        ]
-                    import_details_list.append(import_details)
-                    graph.remove_import(importer=lower_layer_module, imported=imported_module)
-        return import_details_list
+    TODO: Replace with itertools.pairwise once on Python 3.10.
+    """
+    a, b = itertools.tee(iterable)
+    next(b, None)
+    return zip(a, b)